http://bugs.osportfolio.org/jira/browse/OSP-1580 http://bugs.osportfolio.org/jira/browse/OSP-1591 http://bugs.osportfolio.org/jira/browse/OSP-1569 http://bugs.osportfolio.org/jira/browse/OSP-1575 http://bugs.osportfolio.org/jira/browse/OSP-1578 http://bugs.osportfolio.org/jira/browse/OSP-1581 Fixing up some of the Sequential Wizard stuff that was causing problems all over. Clearing out some session variables for the evaluation part of wizards.
git-svn-id: https://source.sakaiproject.org/svn/osp/trunk@11482 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ wizard/tool/src/java/org/theospi/portfolio/wizard/tool/WizardTool.java
@@ -527,7 +527,7 @@ public class WizardTool extends BuilderTool {
             getCurrent().getBase().getType())) {
          session.setAttribute(WizardPageHelper.SEQUENTIAL_WIZARD_PAGES, pages);
          session.setAttribute(WizardPageHelper.SEQUENTIAL_WIZARD_CURRENT_STEP,
-               new Integer(1));
+               new Integer(0));
          redirectAddress = "osp.wizard.page.helper/sequentialWizardPage.osp";
          map.put("finishSeqWizard", LIST_PAGE);
       }
